setup jwt for signup
